refactored db
