+++ app.js
@@ -1,83 +1,61 @@
 const express = require("express");
 const app = express();
-const mongoose = require("mongoose");
-const bodyParser = require("body-parser");
-
-//DB Connection
-mongoose
-    .connect("mongodb://localhost:27017/staycalmDB", {
-        useNewUrlParser: true,
-        useUnifiedTopology: true,
-    })
-    .then(() => console.log("Connected to DB!"))
-    .catch((error) => console.log(error.message));
-
-//SCHEMA SETUP
-const listSchema = new mongoose.Schema({
-    name: String,
-    image: String,
-    city: String,
-    type: String,
-    roomNumber: Number,
-});
-
-// SCHEMA compiled to module
-const List = mongoose.model("List", listSchema);
+const connectDB = require("./models/connection");
 
 app.use(bodyParser.urlencoded({ extended: true }));
 app.set("view engine", "ejs");
 
 app.get("/staycalm", (req, res) => {
-    res.send("StayCalm Project Server started");
+  res.send("StayCalm Project Server started");
 });
 
 app.get("/", (req, res) => {
-    res.render("landing");
+  res.render("landing");
 });
 
 app.get("/listings", (req, res) => {
-    //GET all lists from DB
-    List.find({}, (err, allLists) => {
-        if (err) {
-            console.log(err);
-        } else {
-            res.render("listings", { listings: allLists });
-        }
-    });
+  //GET all lists from DB
+  List.find({}, (err, allLists) => {
+    if (err) {
+      console.log(err);
+    } else {
+      res.render("listings", { listings: allLists });
+    }
+  });
 });
 
 app.post("/listings", (req, res) => {
-    var email = req.body.email;
-    var password = req.body.password;
-    var address = req.body.address;
-    var address2 = req.body.address2;
-    var city = req.body.city;
-    var state = req.body.state;
-    var zip = req.body.zip;
-    var newListing = {
-        email: email,
-        password: password,
-        address: address,
-        address2: address2,
-        city: city,
-        state: state,
-        zip: zip,
-    };
-    List.create(newListing, (err, newlyCreated) => {
-        if (err) {
-            console.log(err);
-        } else {
-            res.redirect("/listings");
-        }
-    });
+  var email = req.body.email;
+  var password = req.body.password;
+  var address = req.body.address;
+  var address2 = req.body.address2;
+  var city = req.body.city;
+  var state = req.body.state;
+  var zip = req.body.zip;
+  var newListing = {
+    email: email,
+    password: password,
+    address: address,
+    address2: address2,
+    city: city,
+    state: state,
+    zip: zip,
+  };
+  List.create(newListing, (err, newlyCreated) => {
+    if (err) {
+      console.log(err);
+    } else {
+      res.redirect("/listings");
+    }
+  });
 });
 
 app.get("/listings/new", (req, res) => {
-    res.render("new.ejs");
+  res.render("new.ejs");
 });
 
 app.listen(5555, () => {
-    console.log("TO STAY CALM YOU NEED STAYCALM");
+  console.log("TO STAY CALM YOU NEED STAYCALM");
 });
 
 // List.find({}, function (err, lists) {
